Remove GBXML DocumentBuilder from workflow for pushing Environment objects to XML
