+++ XML_Adapter/XML_Adapter.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -149,7 +149,6 @@
     <Compile Include="Convert\KML\ToKML.cs" />
     <Compile Include="Convert\KML\ToLatLon.cs" />
     <Compile Include="CRUD\Bluebeam\ReadBluebeam.cs" />
-    <Compile Include="CRUD\Create.cs" />
     <Compile Include="CRUD\CSProject\CreateCSProject.cs" />
     <Compile Include="CRUD\CSProject\ReadCSProject.cs" />
     <Compile Include="CRUD\Default\CreateDefault.cs" />
@@ -242,9 +241,6 @@
       <Name>XML_oM</Name>
       <Private>False</Private>
     </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="Versioning_42.json" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
