+++ XML_Engine/Versioning_63.json
@@ -1,0 +1,7 @@
+﻿{
+  "MessageForDeleted": {
+    "BH.Engine.Adapters.XML.Create.DocumentBuilder(System.Collections.Generic.List<BH.oM.Base.IBHoMObject>)": "The GBXML Document Builder has been retired in favour of allowing objects to be pushed to XML in a manner that is aligned with other BHoM adapters. Please plug your BHoM objects directly into the Push component to continue your workflow. Ensure the XML Schema is set to GBXML in your configuration settings. Please reach out to the development team via https://github.com/BHoM/XML_Toolkit/issues if you encounter any issues with this.",
+    "BH.Engine.Adapters.XML.Create.DocumentBuilder(System.Collections.Generic.List<BH.oM.Environment.Elements.Building>, System.Collections.Generic.List<System.Collections.Generic.List<BH.oM.Environment.Elements.Panel>>, System.Collections.Generic.List<BH.oM.Environment.Elements.Panel>, System.Collections.Generic.List<BH.oM.Spatial.SettingOut.Level>, System.Collections.Generic.List<BH.oM.Environment.Elements.Panel>)": "The GBXML Document Builder has been retired in favour of allowing objects to be pushed to XML in a manner that is aligned with other BHoM adapters. Please plug your BHoM objects directly into the Push component to continue your workflow. Ensure the XML Schema is set to GBXML in your configuration settings. Please reach out to the development team via https://github.com/BHoM/XML_Toolkit/issues if you encounter any issues with this.",
+    "BH.oM.Adapters.XML.GBXMLDocumentBuilder": "The GBXML Document Builder has been retired in favour of allowing objects to be pushed to XML in a manner that is aligned with other BHoM adapters. Please plug your BHoM objects directly into the Push component to continue your workflow. Ensure the XML Schema is set to GBXML in your configuration settings. Please reach out to the development team via https://github.com/BHoM/XML_Toolkit/issues if you encounter any issues with this."
+  }
+}
